scenario 3 implementer

+++ src/main/java/com/sqli/isc/iut/courses/cucumber/Man.java
@@ -4,12 +4,12 @@ public class Man {
 
     private String name;
     private int numberOfDrinks;
-    private double bill;
+    private int bill;
 
     public Man(String name) {
         this.name = name;
         this.numberOfDrinks = 0;
-        this.bill = 0.0;
+        this.bill = 0 ;
     }
 
     public String getName() {
@@ -24,11 +24,11 @@ public class Man {
         return bill;
     }
 
-    public void addToBill(double amount) {
+    public void addToBill(int amount) {
         bill += amount;
     }
 
-    public void setBill(double amount) {
+    public void setBill(int amount) {
         bill = amount;
     }
 
@@ -41,7 +41,7 @@ public class Man {
     }
 
     public void payBill() {
-        bill = 0.0;
+        bill = 0 ;
     }
 
     public boolean isHappy(int numberOfDrinks) {
